% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_overall_metrics}
\alias{calc_overall_metrics}
\title{Overall multiclass classification metrics}
\usage{
calc_overall_metrics(y_true, y_pred, prob, classes)
}
\arguments{
\item{y_true}{True labels.}

\item{y_pred}{Predicted labels.}

\item{prob}{Matrix of predicted probabilities.}

\item{classes}{Character vector of class levels.}
}
\value{
One-row data frame of overall metrics.
}
\description{
Overall multiclass classification metrics
}
