% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_scmr.R
\name{predict_scmr}
\alias{predict_scmr}
\title{Predict from an SCMR package model}
\usage{
predict_scmr(object, ...)
}
\arguments{
\item{object}{A fitted SCMR object.}

\item{...}{Arguments passed to predict().}
}
\value{
Predicted probabilities, classes, or memberships.
}
\description{
Predict from an SCMR package model
}
