% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{read_membership_compact}
\alias{read_membership_compact}
\title{Read all compact membership files from a directory}
\usage{
read_membership_compact(compact_dir)
}
\arguments{
\item{compact_dir}{Directory containing compact membership RDS files.}
}
\value{
Combined long-format membership data frame.
}
\description{
Read all compact membership files from a directory
}
