% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmr-package.R
\docType{package}
\name{scmr-package}
\alias{scmr}
\alias{scmr-package}
\title{Spatially Clustered Multinomial Regression}
\description{
Global, spatially clustered, and fixed-membership multinomial models with
unpenalized, ridge, lasso, and elastic-net fitting. See \code{\link[=fit_scmr]{fit_scmr()}} for fitting,
\code{\link[=scmr_control]{scmr_control()}} for algorithm settings, and \code{\link[=simulate_scmr_data]{simulate_scmr_data()}} for
synthetic data. The article study is distributed in \verb{simulation/} and
\verb{scripts/} within the installed package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/byususen/scmr}
  \item Report bugs at \url{https://github.com/byususen/scmr/issues}
}

}
\author{
\strong{Maintainer}: Bayu Suseno \email{byususen@users.noreply.github.com}

Other contributors:
\itemize{
  \item SCMR study contributors [contributor]
}

}
\keyword{internal}
